pyproject.toml
[tool.poetry]
name = "infared-human-tracking"
version = "0.1.0"
description = ""
authors = ["Edward <edeetee@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
scikit-learn = "^1.5.1"
mido = "^1.3.2"
python-rtmidi = "^1.5.8"
matplotlib = "^3.9.2"
paramiko = "^3.4.1"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
